net/os-carp-vip-dhcp: split codec + capture backends into leasekeeper/
WIP module split: extract the raw wire codec + embedded BPF filter (codec), the
two capture backends with their own optional-import guards (capture_scapy owns
the scapy guard, capture_bpf the fcntl guard), and the backend registry
(capture). _deliver moves to wire (the lowest shared layer) to keep the backend
imports acyclic. lease_keeper.py re-exports the moved names transitionally so
the tests' lk.* access stays green; those move to importing from the package in
the final commit. 123 tests green.
