+++ net/os-carp-vip-dhcp/src/opnsense/scripts/OPNsense/CarpVipDhcp/lease_keeper.py
@@ -74,36 +74,6 @@ from dataclasses import dataclass
 from logging.handlers import RotatingFileHandler
 from typing import Any
 
-# The raw /dev/bpf backend drives its ioctls through fcntl, which does not
-# exist on non-POSIX development hosts. Import it defensively (same pattern as
-# scapy below) so the module still loads there; BpfCapture.start() reports the
-# missing module at runtime instead.
-try:
-    import fcntl as _fcntl_module
-    # Typed Any: the POSIX-only stubs would flag every ioctl call on the
-    # non-POSIX hosts this guard exists for.
-    fcntl: Any = _fcntl_module
-except ImportError:
-    fcntl = None  # pylint: disable=invalid-name
-
-# Scapy is the one heavy third-party dependency (needed only by the default
-# scapy capture backend) and the usual suspect when the daemon won't start
-# (missing after an upgrade, ABI mismatch, partial install).
-# Import it defensively: a bare module-level import that throws would kill the
-# process before any log handler exists -> a silent non-start with the traceback
-# going to daemon(8)'s /dev/null. Capture it instead and let main() log it once
-# logging is configured (only when the scapy backend is actually selected).
-try:
-    from scapy.all import (  # type: ignore  # pylint: disable=import-error
-        ARP, Ether, IP, UDP, BOOTP, DHCP, AsyncSniffer, sendp)
-    _SCAPY_IMPORT_ERROR = None
-except Exception as _scapy_exc:   # ImportError, or scapy's own import-time failures
-    _SCAPY_IMPORT_ERROR = _scapy_exc
-    # Bind the names so the module still loads (main() exits with the captured
-    # error before any of them is used). Typed Any so type checkers do not
-    # flag every scapy call site as calling None.
-    _MISSING: Any = None
-    ARP = Ether = IP = UDP = BOOTP = DHCP = AsyncSniffer = sendp = _MISSING  # pylint: disable=invalid-name
 
 LOG = logging.getLogger("lease-keeper")
 
@@ -125,628 +95,27 @@ from leasekeeper.wire import (
     ArpFrame, BootpFrame, DhcpReply, SNIFFER_FILTER, _dhcp_options, _fmt_reply,
     _msg_text, _parse_reply)
 
+from leasekeeper.capture import CAPTURE_BACKENDS
+from leasekeeper.capture_scapy import _SCAPY_IMPORT_ERROR
+# Transitional re-exports: the tests reach these via the lease_keeper module
+# (lk.*); they move to importing from leasekeeper.* directly in the final
+# commit of the split, when these lines go away.
+from leasekeeper.capture_scapy import (  # noqa: F401  pylint: disable=unused-import
+    ARP, BOOTP, DHCP, Ether, IP, UDP, AsyncSniffer, sendp, ScapyCapture)
+from leasekeeper.capture_bpf import BpfCapture  # noqa: F401  pylint: disable=unused-import
+from leasekeeper.codec import (  # noqa: F401  pylint: disable=unused-import
+    BOOTP_HDR_LEN, BOOTP_MIN_PAYLOAD, BPF_ALIGNMENT, BPF_HDR_FIXED, DHCP_MAGIC,
+    ETHERTYPE_ARP, ETHERTYPE_IPV4, ETHER_MIN_FRAME, IPPROTO_UDP, _BPF_FILTER,
+    _bpf_frames, _decode_arp, _decode_dhcp_options, _decode_ipv4_bootp,
+    _encode_arp_request, _encode_bootp_request, _encode_dhcp_options,
+    _encode_ether, _encode_ipv4_udp, _inet_checksum, _ip4)
 
-# ---- raw wire codec (bpf backend) ----
-# Hand encoders/decoders for exactly the frames the keeper exchanges, so the
-# bpf backend needs no packet library. The decoders parse untrusted WAN input:
-# every access is bounds-checked and a malformed frame decodes to None
-# (dropped) rather than raising.
 
-ETHERTYPE_IPV4 = 0x0800
-ETHERTYPE_ARP = 0x0806
-ETHER_MIN_FRAME = 60         # minimum Ethernet frame (without FCS); short ARP frames are padded
-DHCP_MAGIC = b"\x63\x82\x53\x63"   # RFC 2131 options magic cookie
-BOOTP_HDR_LEN = 236          # fixed BOOTP header before the magic cookie
-BOOTP_MIN_PAYLOAD = 300      # RFC 1542 4.1 minimum BOOTP message; reference clients pad to it
-DHCP_OPT_PAD = 0
-DHCP_OPT_END = 255
-IPPROTO_UDP = 17
-IPV4_TTL = 64                # conventional default TTL (BSD/Linux/scapy send 64)
 
-# DHCP message types the keeper SENDS, by the names the option lists carry
-# (the inverse, code -> display name, is MTYPE_NAMES above).
-_MTYPE_CODES = {"discover": 1, "request": 3, "release": 7}
 
 
-def _ip4(ip):
-    """Dotted-quad string (None -> 0.0.0.0) -> 4 raw bytes."""
-    return ipaddress.IPv4Address(ip or "0.0.0.0").packed
 
-
-def _ip4_str(raw):
-    """4 raw bytes -> dotted-quad string."""
-    return str(ipaddress.IPv4Address(bytes(raw)))
-
-
-def _opt_text(v):
-    """Text-ish option value -> bytes (client-id arrives pre-encoded)."""
-    return v if isinstance(v, bytes) else str(v).encode()
-
-
-# Outbound option encoders: option name -> (wire code, value encoder) for
-# exactly the options the keeper sends. These names ARE the keeper's option
-# vocabulary (BootpFrame.options and _dhcp_options use them).
-_OPT_ENCODERS = {
-    "message-type": (53, lambda v: bytes([_MTYPE_CODES[v]])),
-    "param_req_list": (55, bytes),        # list of option codes
-    "requested_addr": (50, _ip4),
-    "server_id": (54, _ip4),
-    "hostname": (12, _opt_text),
-    "vendor_class_id": (60, _opt_text),
-    "client_id": (61, _opt_text),
-}
-
-
-def _encode_dhcp_options(options):
-    """A scapy-style option list (name/value tuples + "end") -> the raw options
-    field. Options with a None value are dropped (a broadcast RELEASE carries
-    no server-id); the end option is always appended."""
-    out = bytearray()
-    for o in options:
-        if not isinstance(o, tuple) or o[1] is None:
-            continue   # the "end" marker (appended below) or a valueless option
-        code, encode = _OPT_ENCODERS[o[0]]
-        raw = encode(o[1])
-        out += bytes([code, len(raw)]) + raw
-    out.append(DHCP_OPT_END)
-    return bytes(out)
-
-
-def _encode_bootp_request(chaddr, xid, ciaddr, flags, options):
-    """A BOOTREQUEST payload: fixed BOOTP header + cookie + options, padded to
-    the RFC 1542 minimum message size (some servers drop shorter ones)."""
-    hdr = struct.pack("!4BIHH", 1, 1, 6, 0, xid, 0, flags)   # op htype hlen hops xid secs flags
-    hdr += _ip4(ciaddr) + b"\x00" * 12                       # ciaddr, then yiaddr/siaddr/giaddr zero
-    hdr += chaddr.ljust(16, b"\x00")[:16]                    # chaddr field (16 bytes)
-    hdr += b"\x00" * 192                                     # sname (64) + file (128), unused
-    payload = hdr + DHCP_MAGIC + _encode_dhcp_options(options)
-    return payload.ljust(BOOTP_MIN_PAYLOAD, b"\x00")
-
-
-def _inet_checksum(data):
-    """RFC 1071 ones-complement sum over 16-bit words (odd length zero-padded)."""
-    if len(data) % 2:
-        data += b"\x00"
-    total = sum(int.from_bytes(data[i:i + 2], "big") for i in range(0, len(data), 2))
-    while total >> 16:
-        total = (total & 0xFFFF) + (total >> 16)
-    return ~total & 0xFFFF
-
-
-def _encode_ipv4_udp(src, dst, sport, dport, payload):
-    """An IPv4+UDP datagram around payload, checksums computed. A UDP checksum
-    of 0 means "none sent", so a sum that works out to 0 transmits as 0xFFFF
-    (RFC 768)."""
-    udp_len = 8 + len(payload)
-    pseudo = _ip4(src) + _ip4(dst) + struct.pack("!BBH", 0, IPPROTO_UDP, udp_len)
-    udp_hdr = struct.pack("!4H", sport, dport, udp_len, 0)
-    cksum = _inet_checksum(pseudo + udp_hdr + payload) or 0xFFFF
-    udp_hdr = struct.pack("!4H", sport, dport, udp_len, cksum)
-    # version/ihl, tos, total, id, flags/frag, ttl, proto, checksum (zeroed, then patched)
-    ip_hdr = struct.pack("!BBHHHBBH", 0x45, 0, 20 + udp_len, 0, 0, IPV4_TTL, IPPROTO_UDP, 0) + _ip4(src) + _ip4(dst)
-    ip_hdr = ip_hdr[:10] + _inet_checksum(ip_hdr).to_bytes(2, "big") + ip_hdr[12:]
-    return ip_hdr + udp_hdr + payload
-
-
-def _encode_ether(dst, src, ethertype, payload):
-    """An Ethernet II frame."""
-    return mac2raw(dst) + mac2raw(src) + ethertype.to_bytes(2, "big") + payload
-
-
-# The fixed Ethernet/IPv4 ARP header prefix (htype/ptype/hlen/plen), shared by
-# the encoder and the decoder so the two cannot drift.
-_ARP_ETH_IPV4 = struct.pack("!HHBB", 1, ETHERTYPE_IPV4, 6, 4)
-
-
-def _encode_arp_request(hwsrc, psrc, pdst):
-    """An ARP who-has pdst tell psrc with sender hardware hwsrc (the nudge
-    frame; the shaping rationale lives at the ArpNudge call site)."""
-    return (_ARP_ETH_IPV4 + struct.pack("!H", 1)
-            + mac2raw(hwsrc) + _ip4(psrc) + mac2raw(ETHER_ZERO) + _ip4(pdst))
-
-
-def _u32(v):
-    """4-byte big-endian option value -> int (a shorter value is malformed)."""
-    if len(v) < 4:
-        raise ValueError("short integer option")
-    return int.from_bytes(v[:4], "big")
-
-
-def _first_ip(v):
-    """First IPv4 address in an option value (option 3 may list several)."""
-    if len(v) < 4:
-        raise ValueError("short address option")
-    return _ip4_str(v[:4])
-
-
-# Inbound option decoders: wire code -> (option name, value decoder) for
-# exactly the options _parse_reply acts on; everything else is skipped unread
-# (untrusted input, narrow surface).
-_OPT_DECODERS = {
-    1: ("subnet_mask", _first_ip),
-    3: ("router", _first_ip),
-    51: ("lease_time", _u32),
-    53: ("message-type", lambda v: v[0]),
-    54: ("server_id", _first_ip),
-    56: ("message", bytes),
-    58: ("renewal_time", _u32),
-    59: ("rebinding_time", _u32),
-}
-
-
-def _decode_dhcp_options(data):
-    """Bounds-checked TLV walk over the options field of an untrusted reply.
-    Unknown options are skipped without decoding; any truncation ends the walk
-    with what was parsed so far. Never raises on malformed input.
-
-    Option overload (option 52), where a server continues its options into the
-    BOOTP sname/file fields, is deliberately NOT handled: with the keeper's
-    6-entry parameter request list every reply fits the options field many
-    times over, so no server the keeper talks to overloads."""
-    options = []
-    pos = 0
-    while pos < len(data):
-        code = data[pos]
-        if code == DHCP_OPT_PAD:        # a single padding byte, no length/value
-            pos += 1
-            continue
-        if code == DHCP_OPT_END:        # explicit end of options
-            break
-        if pos + 2 > len(data):         # the length byte itself is missing
-            break
-        length = data[pos + 1]
-        value = data[pos + 2:pos + 2 + length]
-        if len(value) < length:         # the value runs past the end of the buffer
-            break
-        decoder = _OPT_DECODERS.get(code)
-        if decoder is not None:
-            name, decode_value = decoder
-            try:
-                options.append((name, decode_value(value)))
-            except (ValueError, IndexError):
-                pass                    # malformed value in a known option: skip, keep walking
-        pos += 2 + length
-    return options
-
-
-def _decode_arp(pkt):
-    """Raw ARP payload -> ArpFrame, or None unless it is a well-formed
-    Ethernet/IPv4 ARP."""
-    if len(pkt) < 28 or pkt[:6] != _ARP_ETH_IPV4:
-        return None
-    return ArpFrame(int.from_bytes(pkt[6:8], "big"), _ip4_str(pkt[14:18]), _ip4_str(pkt[24:28]))
-
-
-def _decode_ipv4_bootp(pkt):
-    """Raw IPv4 payload of an Ethernet frame -> BootpFrame, or None unless it
-    is an unfragmented UDP datagram carrying a plausible BOOTP message with
-    the DHCP cookie. Every bound is checked (untrusted WAN input)."""
-    if len(pkt) < 20 or pkt[0] >> 4 != 4:
-        return None
-    ihl = (pkt[0] & 0x0F) * 4
-    total = int.from_bytes(pkt[2:4], "big")
-    if ihl < 20 or total < ihl + 8 or len(pkt) < ihl + 8:
-        return None
-    is_udp = pkt[9] == IPPROTO_UDP
-    is_unfragmented = (int.from_bytes(pkt[6:8], "big") & 0x3FFF) == 0   # no MF bit, no frag offset
-    if not (is_udp and is_unfragmented):
-        return None
-    # Bound the BOOTP slice by the UDP length too, not just the IP total: a
-    # datagram whose UDP length is shorter than the IP payload would otherwise
-    # let trailing bytes be parsed as DHCP options. Trust the smallest of the
-    # three lengths (untrusted input); a UDP length shorter than its own header
-    # is malformed.
-    udp_len = int.from_bytes(pkt[ihl + 4:ihl + 6], "big")
-    if udp_len < 8:
-        return None
-    end = min(total, ihl + udp_len, len(pkt))
-    bootp = pkt[ihl + 8:end]                        # UDP payload, minus any link padding
-    if len(bootp) < BOOTP_HDR_LEN + 4 or bootp[BOOTP_HDR_LEN:BOOTP_HDR_LEN + 4] != DHCP_MAGIC:
-        return None
-    # The options walk only pays off for server replies; on a shared segment
-    # the filter also passes other clients' broadcast BOOTREQUESTs, which the
-    # keeper drops on op anyway -- skip their TLV walk in this hot path.
-    op = bootp[0]
-    return BootpFrame(op=op,
-                      xid=int.from_bytes(bootp[4:8], "big"),
-                      yiaddr=_ip4_str(bootp[16:20]),
-                      chaddr=bytes(bootp[28:44]),
-                      giaddr=_ip4_str(bootp[24:28]),
-                      options=_decode_dhcp_options(bootp[BOOTP_HDR_LEN + 4:]) if op == BOOTREPLY else [])
-
-
-# ---- /dev/bpf plumbing (bpf backend) ----
-# FreeBSD ioctl codes for the LP64 platforms OPNsense ships on (amd64/aarch64),
-# precomputed from net/bpf.h so no C headers are needed at runtime.
-BIOCGBLEN = 0x40044266       # _IOR('B', 102, u_int): kernel capture buffer size
-BIOCSETF = 0x80104267        # _IOW('B', 103, struct bpf_program): attach the filter
-BIOCPROMISC = 0x20004269     # _IO('B', 105): promiscuous mode
-BIOCGDLT = 0x4004426A        # _IOR('B', 106, u_int): the interface's data-link type
-BIOCSETIF = 0x8020426C       # _IOW('B', 108, struct ifreq): bind to the interface
-BIOCIMMEDIATE = 0x80044270   # _IOW('B', 112, u_int): deliver per packet, not per full buffer
-BIOCSHDRCMPLT = 0x80044275   # _IOW('B', 117, u_int): we supply the Ethernet source MAC
-DLT_EN10MB = 1               # Ethernet: the only link type this codec's offsets assume
-BPF_ALIGNMENT = 8            # capture records align to sizeof(long)
-BPF_HDR_FIXED = 26           # bh_tstamp(16) + bh_caplen(4) + bh_datalen(4) + bh_hdrlen(2)
-
-# SNIFFER_FILTER compiled to classic-BPF opcodes, embedded so the daemon needs
-# no runtime filter compiler. MUST stay in lockstep with SNIFFER_FILTER;
-# regenerate on any FreeBSD/OPNsense host with:
-#   tcpdump -i <ethernet-iface> -dd '(udp and (port 67 or port 68)) or (arp and arp[6:2] = 2)'
-# The trailing comment on each row is the `tcpdump -d` mnemonic so the table can
-# be audited by eye against the filter string without a FreeBSD box; jump
-# targets are absolute instruction indices (tcpdump's relative jt/jf + here+1).
-# A bench test (testbench repo) asserts `tcpdump -ddd SNIFFER_FILTER` still
-# equals this table, turning the lockstep requirement into an enforced invariant.
-_BPF_FILTER = (
-    (0x28, 0, 0, 0x0000000C),   # 00 ldh  [12]                 ; ethertype
-    (0x15, 0, 10, 0x00000800),  # 01 jeq  #0x800  -> 02, ->12  ; IPv4?
-    (0x30, 0, 0, 0x00000017),   # 02 ldb  [23]                 ; IPv4 proto
-    (0x15, 0, 21, 0x00000011),  # 03 jeq  #17     -> 04, ->25  ; UDP?
-    (0x28, 0, 0, 0x00000014),   # 04 ldh  [20]                 ; flags+frag
-    (0x45, 19, 0, 0x00001FFF),  # 05 jset #0x1fff ->25, -> 06  ; fragment? drop
-    (0xB1, 0, 0, 0x0000000E),   # 06 ldxb 4*([14]&0xf)         ; X = IP hdr len
-    (0x48, 0, 0, 0x0000000E),   # 07 ldh  [x+14]               ; UDP src port
-    (0x15, 15, 0, 0x00000043),  # 08 jeq  #67     ->24, -> 09  ; sport 67? accept
-    (0x15, 14, 0, 0x00000044),  # 09 jeq  #68     ->24, -> 10  ; sport 68? accept
-    (0x48, 0, 0, 0x00000010),   # 10 ldh  [x+16]               ; UDP dst port
-    (0x15, 12, 8, 0x00000043),  # 11 jeq  #67     ->24, ->20   ; dport 67? accept
-    (0x15, 0, 8, 0x000086DD),   # 12 jeq  #0x86dd -> 13, ->21  ; IPv6? (else ARP)
-    (0x30, 0, 0, 0x00000014),   # 13 ldb  [20]                 ; IPv6 next header
-    (0x15, 0, 10, 0x00000011),  # 14 jeq  #17     -> 15, ->25  ; UDP?
-    (0x28, 0, 0, 0x00000036),   # 15 ldh  [54]                 ; UDP src port (14+40)
-    (0x15, 7, 0, 0x00000043),   # 16 jeq  #67     ->24, -> 17  ; sport 67? accept
-    (0x15, 6, 0, 0x00000044),   # 17 jeq  #68     ->24, -> 18  ; sport 68? accept
-    (0x28, 0, 0, 0x00000038),   # 18 ldh  [56]                 ; UDP dst port
-    (0x15, 4, 0, 0x00000043),   # 19 jeq  #67     ->24, ->20   ; dport 67? accept
-    (0x15, 3, 4, 0x00000044),   # 20 jeq  #68     ->24, ->25   ; dport 68? accept
-    (0x15, 0, 3, 0x00000806),   # 21 jeq  #0x806  -> 22, ->25  ; ARP?
-    (0x28, 0, 0, 0x00000014),   # 22 ldh  [20]                 ; ARP opcode
-    (0x15, 0, 1, 0x00000002),   # 23 jeq  #2      ->24, ->25   ; is-at reply?
-    (0x6, 0, 0, 0x00040000),    # 24 ret  #262144              ; ACCEPT (snap len)
-    (0x6, 0, 0, 0x00000000),    # 25 ret  #0                   ; DROP
-)
-
-
-def _bpf_align(nbytes):
-    """Round a record length up to the BPF record alignment (sizeof(long))."""
-    return (nbytes + BPF_ALIGNMENT - 1) & ~(BPF_ALIGNMENT - 1)
-
-
-def _bpf_frames(data):
-    """Yield each captured Ethernet frame from a raw BPF read buffer.
-
-    One read(2) can return several packets back to back, each prefixed by a
-    struct bpf_hdr and padded to the record alignment. bh_tstamp occupies the
-    first 16 bytes; bh_caplen, bh_datalen and bh_hdrlen follow. A record whose
-    header or length is inconsistent ends the walk rather than risk misparsing
-    the rest of the buffer (it is only as trustworthy as the kernel handed it
-    over)."""
-    pos = 0
-    while pos + BPF_HDR_FIXED <= len(data):
-        # bh_caplen (u_int), bh_datalen (u_int), bh_hdrlen (u_short), in host
-        # byte order, immediately after the 16-byte bh_tstamp.
-        caplen, _datalen, hdrlen = struct.unpack_from("=IIH", data, pos + 16)
-        record_end = pos + hdrlen + caplen
-        if hdrlen < BPF_HDR_FIXED or record_end > len(data):
-            break
-        yield data[pos + hdrlen:record_end]
-        pos += _bpf_align(hdrlen + caplen)
-
-
-# ---- capture backends ----
-# Both backends expose the same surface: start/stop/alive for the capture
-# side (decoded BootpFrame/ArpFrame handed to the constructor callbacks on
-# the capture thread) and send_dhcp/send_arp_request for the send side.
-
-def _deliver(handler, frame):
-    """Run a keeper frame callback under its own guard: a failure in there is
-    a handler bug, not a parse error, and must neither kill the capture
-    thread nor be mislabelled as malformed input."""
-    if frame is None:
-        return
-    try:
-        handler(frame)
-    except Exception as e:
-        LOG.debug("frame handler error: %s", e)
-
-
-class ScapyCapture:
-    """Capture/send via scapy (AsyncSniffer + sendp): the default backend.
-    Decodes scapy packets into the same neutral frames as the bpf backend, so
-    the rest of the keeper never touches a packet object. Outbound option
-    lists are relayed to scapy verbatim, which is why the keeper's option
-    vocabulary must stay scapy-compatible."""
-
-    def __init__(self, iface, promisc, on_bootp, on_arp):
-        self.iface = iface
-        self.promisc = promisc
-        self._on_bootp = on_bootp
-        self._on_arp = on_arp
-        self._sniffer = None
-
-    def start(self):
-        """(Re)start the sniffer. False on failure (the caller retries)."""
-        try:
-            self.stop()
-            # Non-promiscuous by default: the BOOTP broadcast flag makes the
-            # server broadcast OFFER/ACK, and the gateway's unicast ARP reply to a
-            # nudge reaches us because the CARP master already accepts the VIP's
-            # virtual MAC. promisc is the opt-in fallback for NICs that
-            # drop non-primary unicast (widens capture -- warned at startup).
-            self._sniffer = AsyncSniffer(
-                iface=self.iface, filter=SNIFFER_FILTER,
-                prn=self._on_packet, store=0, promisc=self.promisc)
-            self._sniffer.start()
-            return True
-        except Exception as e:
-            LOG.error("DHCP-reply sniffer start failed: %s", e)
-            return False
-
-    def stop(self):
-        """Best-effort capture stop (scapy may raise if it never ran).
-
-        Join with a bound: scapy's own stop() joins the sniffer thread with no
-        timeout, so a sniffer that fails to break its loop would hang the main
-        thread here -- and because this runs from _ensure_sniffer mid-sequence,
-        that would freeze the heartbeat and trip a false CARP demotion. Ask it
-        not to join, then join the thread ourselves with a ceiling."""
-        try:
-            if self._sniffer:
-                self._sniffer.stop(join=False)
-                thread = getattr(self._sniffer, "thread", None)
-                if thread is not None:
-                    thread.join(timeout=2)
-        except Exception:
-            pass
-
-    def alive(self):
-        """True while the sniffer thread is running."""
-        thread = getattr(self._sniffer, "thread", None)
-        return thread is not None and thread.is_alive()
-
-    def _on_packet(self, p):
-        """Sniffer callback: scapy packet -> neutral frame -> keeper callback
-        (via _deliver, so a handler failure is labelled as such). A parse
-        error in the untrusted input is dropped (debug-logged)."""
-        frame = handler = None
-        try:
-            if p.haslayer(BOOTP) and p.haslayer(DHCP):
-                b = p[BOOTP]
-                try:
-                    chaddr = bytes(getattr(b, "chaddr", b"") or b"")
-                except (TypeError, ValueError):
-                    chaddr = b""
-                frame = BootpFrame(b.op, b.xid, b.yiaddr, chaddr,
-                                   getattr(b, "giaddr", None), p[DHCP].options)
-                handler = self._on_bootp
-            elif p.haslayer(ARP):
-                arp = p[ARP]
-                frame, handler = ArpFrame(arp.op, arp.psrc, arp.pdst), self._on_arp
-        except Exception as e:
-            LOG.debug("sniffed packet parse error: %s", e)
-            return
-        _deliver(handler, frame)
-
-    # The DHCP wire tuple: one parameter per field that goes on the wire.
-    def send_dhcp(self, *, eth_src, ip_src, ip_dst, chaddr,  # pylint: disable=too-many-arguments
-                  xid, ciaddr, flags, options):
-        """Broadcast one DHCP client message as scapy layers."""
-        sendp(Ether(src=eth_src, dst=ETHER_BROADCAST) /
-              IP(src=ip_src, dst=ip_dst) /
-              UDP(sport=DHCP_CLIENT_PORT, dport=DHCP_SERVER_PORT) /
-              BOOTP(chaddr=chaddr, xid=xid, ciaddr=ciaddr, flags=flags) /
-              DHCP(options=options),
-              iface=self.iface, verbose=0)
-
-    def send_arp_request(self, hwsrc, psrc, pdst):
-        """Broadcast an ARP who-has pdst tell psrc from hwsrc."""
-        sendp(Ether(src=hwsrc, dst=ETHER_BROADCAST) /
-              ARP(op=1, hwsrc=hwsrc, psrc=psrc,
-                  hwdst=ETHER_ZERO, pdst=pdst),
-              iface=self.iface, verbose=0)
-
-
-class BpfCapture:  # pylint: disable=too-many-instance-attributes
-    """Capture/send on a raw /dev/bpf descriptor -- no packet library. A
-    reader thread walks the BPF buffer and hands decoded neutral frames to
-    the same callbacks the scapy backend feeds. FreeBSD-only (OPNsense's
-    platform); selected with --capture-backend bpf (experimental).
-
-    Shutdown uses a self-pipe rather than a poll timeout: the reader blocks in
-    select() on both the bpf fd and a wake pipe, and stop() writes one byte to
-    the pipe so the reader returns at once (no periodic wakeups, no up-to-1s
-    stop latency). The stop signal and the wake pipe are created fresh per
-    start(), and the reader owns and closes its own bpf fd on exit, so a reader
-    that outlives its stop() (e.g. stalled in a slow log write) can neither be
-    revived by the next start() nor have its fd number reused underneath it."""
-
-    def __init__(self, iface, promisc, on_bootp, on_arp):
-        self.iface = iface
-        self.promisc = promisc
-        self._on_bootp = on_bootp
-        self._on_arp = on_arp
-        self._fd = None                # the live capture fd, or None when stopped
-        self._buflen = 0               # kernel buffer size, from BIOCGBLEN in _configure
-        self._thread = None            # the current reader thread
-        self._stop_event = None        # set to ask the current reader to exit
-        self._wake_writer = None       # write end of this generation's wake pipe
-
-    def start(self):
-        """(Re)open /dev/bpf, bind + filter it to the interface, and start the
-        reader thread. Returns False on any failure (the caller retries)."""
-        if fcntl is None:
-            LOG.error("bpf backend unavailable: no fcntl module on this platform")
-            return False
-        self.stop()
-        wake_reader, wake_writer = os.pipe()
-        try:
-            fd = os.open("/dev/bpf", os.O_RDWR)
-        except OSError as e:
-            os.close(wake_reader)
-            os.close(wake_writer)
-            LOG.error("bpf capture start failed on %s: %s", self.iface, e)
-            return False
-        try:
-            self._configure(fd)
-        except Exception as e:
-            os.close(fd)
-            os.close(wake_reader)
-            os.close(wake_writer)
-            LOG.error("bpf capture start failed on %s: %s", self.iface, e)
-            return False
-        stop_event = threading.Event()
-        # The reader owns fd and wake_reader and closes them when it exits.
-        self._thread = threading.Thread(
-            target=self._read_loop, args=(fd, wake_reader, stop_event),
-            name="bpf-capture", daemon=True)
-        self._fd = fd
-        self._stop_event = stop_event
-        self._wake_writer = wake_writer
-        self._thread.start()
-        return True
-
-    def _configure(self, fd):
-        """Bind, tune and filter a fresh bpf descriptor.
-
-        BIOCSETIF (bind) has to come first -- it is what libpcap does and what
-        BIOCGDLT needs -- so the filter is attached immediately after, keeping
-        the window in which the descriptor would accept unfiltered traffic to a
-        single ioctl."""
-        fcntl.ioctl(fd, BIOCSETIF, struct.pack("16s16x", self.iface.encode()))
-        # The codec's frame offsets assume Ethernet; a PPPoE/tun WAN would make
-        # both capture and injection meaningless. Fail loudly instead of leaving
-        # only a "no DHCP OFFER" symptom.
-        dlt = struct.unpack("I", fcntl.ioctl(fd, BIOCGDLT, b"\x00" * 4))[0]
-        if dlt != DLT_EN10MB:
-            raise OSError(f"{self.iface} is not Ethernet (bpf data-link type {dlt}); "
-                          "the bpf backend supports Ethernet only")
-        # Attach the capture filter right after the bind (and before the rest of
-        # the tuning) so almost no unfiltered traffic can enter the buffer.
-        program = b"".join(struct.pack("HBBI", *insn) for insn in _BPF_FILTER)
-        program_buf = ctypes.create_string_buffer(program)   # kernel copies it during the ioctl
-        # struct bpf_program is { u_int bf_len; struct bpf_insn *bf_insns; };
-        # "@IQ" gives the native u_int + pointer layout on LP64.
-        fcntl.ioctl(fd, BIOCSETF,
-                    struct.pack("@IQ", len(_BPF_FILTER), ctypes.addressof(program_buf)))
-        # Immediate mode: hand packets over as they arrive; the DHCP exchanges
-        # wait on second-scale timeouts, so buffering a full block is not an option.
-        fcntl.ioctl(fd, BIOCIMMEDIATE, struct.pack("I", 1))
-        # Header-complete: our frames carry the CARP vMAC as the Ethernet source;
-        # without this the kernel would overwrite it with the NIC's own MAC.
-        fcntl.ioctl(fd, BIOCSHDRCMPLT, struct.pack("I", 1))
-        if self.promisc:
-            fcntl.ioctl(fd, BIOCPROMISC)
-        # bpf read(2) calls must request exactly the kernel buffer size.
-        self._buflen = struct.unpack("I", fcntl.ioctl(fd, BIOCGBLEN, b"\x00" * 4))[0]
-
-    def stop(self):
-        """Ask the current reader to exit and wait briefly for it. The reader
-        closes its own fd, so stop() only signals and joins; a reader still
-        alive after the join (stuck in a slow callback) is left to exit on its
-        own -- it holds its own fd, so nothing here can be reused under it."""
-        thread = self._thread
-        stop_event = self._stop_event
-        wake_writer = self._wake_writer
-        self._fd = None
-        self._thread = None
-        self._stop_event = None
-        self._wake_writer = None
-
-        if stop_event is not None:
-            stop_event.set()
-        if wake_writer is not None:
-            try:
-                os.write(wake_writer, b"\x00")   # wake the reader's select() at once
-            except OSError:
-                pass
-            try:
-                os.close(wake_writer)
-            except OSError:
-                pass
-        if thread is not None and thread.is_alive():
-            thread.join(timeout=2)
-            if thread.is_alive():
-                LOG.warning("bpf reader did not exit within 2s -- leaving it to "
-                            "finish; its fd is not reused")
-
-    def alive(self):
-        """True while the descriptor is open and the reader thread runs."""
-        return self._fd is not None and self._thread is not None and self._thread.is_alive()
-
-    def _read_loop(self, fd, wake_reader, stop_event):
-        """Reader thread: block in select() on the bpf fd and the wake pipe;
-        stop() writes to the pipe to end the wait. Owns fd and wake_reader and
-        closes both on exit, so the fd's lifetime ends exactly when this thread
-        does."""
-        try:
-            while not stop_event.is_set():
-                try:
-                    readable, _, _ = select.select([fd, wake_reader], [], [])
-                except OSError:
-                    return
-                if wake_reader in readable:      # stop() rang -- loop condition ends us
-                    continue
-                try:
-                    data = os.read(fd, self._buflen)
-                except (OSError, ValueError):
-                    if not stop_event.is_set():
-                        LOG.warning("bpf read failed -- capture thread exiting")
-                    return
-                for frame in _bpf_frames(data):
-                    self._dispatch(frame)
-        finally:
-            for owned_fd in (fd, wake_reader):
-                try:
-                    os.close(owned_fd)
-                except OSError:
-                    pass
-
-    def _dispatch(self, frame):
-        """Decode one captured Ethernet frame and route it by ethertype to the
-        keeper callback (via _deliver, so a handler failure is labelled as
-        such). A parse error in the untrusted input is dropped (debug-logged)."""
-        decoded = None
-        handler = None
-        try:
-            if len(frame) >= 14:
-                ethertype = int.from_bytes(frame[12:14], "big")
-                if ethertype == ETHERTYPE_ARP:
-                    decoded, handler = _decode_arp(frame[14:]), self._on_arp
-                elif ethertype == ETHERTYPE_IPV4:
-                    decoded, handler = _decode_ipv4_bootp(frame[14:]), self._on_bootp
-        except Exception as e:
-            LOG.debug("bpf frame parse error: %s", e)
-            return
-        _deliver(handler, decoded)
-
-    def _write(self, frame):
-        """Inject one raw Ethernet frame on the interface. Main-thread only (the
-        capture thread never sends), so reading self._fd needs no lock."""
-        fd = self._fd
-        if fd is None:
-            raise OSError("bpf capture not started")
-        os.write(fd, frame)
-
-    # The DHCP wire tuple: one parameter per field that goes on the wire.
-    def send_dhcp(self, *, eth_src, ip_src, ip_dst, chaddr,  # pylint: disable=too-many-arguments
-                  xid, ciaddr, flags, options):
-        """Broadcast one DHCP client message as raw encoded frames."""
-        payload = _encode_bootp_request(chaddr, xid, ciaddr, flags, options)
-        dgram = _encode_ipv4_udp(ip_src, ip_dst, DHCP_CLIENT_PORT, DHCP_SERVER_PORT, payload)
-        self._write(_encode_ether(ETHER_BROADCAST, eth_src, ETHERTYPE_IPV4, dgram))
-
-    def send_arp_request(self, hwsrc, psrc, pdst):
-        """Broadcast an ARP who-has pdst tell psrc from hwsrc."""
-        frame = _encode_ether(ETHER_BROADCAST, hwsrc, ETHERTYPE_ARP,
-                              _encode_arp_request(hwsrc, psrc, pdst))
-        self._write(frame.ljust(ETHER_MIN_FRAME, b"\x00"))   # runt guard: ARP is 42 bytes bare
-
-
-# The capture-backend registry: flag value -> implementation. The argparse
-# choices and Keeper's lookup both read this, so a future backend is added in
 # exactly one place (plus its rc.conf documentation).
-CAPTURE_BACKENDS = {"scapy": ScapyCapture, "bpf": BpfCapture}
